requirements.txt
asgiref==3.3.1
cffi==1.14.5
configobj==5.0.6
cryptography==3.4.6
Django==3.1.7
djangorestframework==3.12.2
email==6.0.0a1
future==0.18.2
more-itertools==8.7.0
Pillow==8.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-gnupg==0.4.6
python-ldap==3.3.1
pytz==2021.1
secrets==1.0.2
six==1.15.0
sqlparse==0.4.1
systematic==4.8.7
